Continue Agro, RAP => Eidos and free WebApp from EidosShell
Move LineReader to Utils, update batch files, "retire" some commented-out code
