+++ shell
@@ -1,2 +1,2 @@
 #!/bin/bash
-sbt 'run-main agro.demo.RAPShell'
+sbt 'run-main eidos.demo.EidosShell'
