+++ src/main/scala/eidos/demo/ExampleGenerator.scala
@@ -1,4 +1,4 @@
-package agro.demo
+package eidos.demo
 
 import java.io.PrintWriter
