+++ src/main/scala/org/clulab/wm/EidosSystem.scala
@@ -9,7 +9,7 @@ import org.clulab.wm.entities.EidosEntityFinder
 import org.clulab.wm.wmutils.FileUtils.{loadDomainParams, loadGradableAdjGroundingFile, readRules}
 
 /**
-  * Handles text processing and information extraction for Agro domain.
+  * A system for text processing and information extraction
   */
 class EidosSystem(
   // The first three are loaded as resources from URLs, thus the leading /
